% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_p.R
\name{adjust_p}
\alias{adjust_p}
\title{Adjust the Pvalue Matrix for Bipartite Network}
\usage{
adjust_p(p, method = "qvalue", ...)
}
\arguments{
\item{p}{The p-value matrix is an object where SNPs are in rows and
phentoypes are in columns. The entries represent the strength of
association between SNP and phenotype. This could be a simple matrix
or data frame.}

\item{method}{A string that specifies what methods will
be used to adjust p-value, which can be "qvalue","lfdr (local FDR)",
or "BH (Benjamini & Hochberg (1995))". The "BH" method is time consuming
for the huge matrix. The default method is "qvalue".}

\item{\dots}{Arguments passed to the \code{qvalue()}.}
}
\value{
A matrix of adjusted p-value
}
\description{
`adjust_p()` : Given a set of p-values, returns p-values adjusted
 using one of several methods.
}
\keyword{GPN}
