% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Adjusted_Tstat}
\alias{Adjusted_Tstat}
\title{Adjusted test statistics calculation}
\usage{
Adjusted_Tstat(x, y, cov = NULL)
}
\arguments{
\item{x}{The individual-level genotypes with dimension n individuals by M SNPs.
It could be only one SNPs with a n by 1 genotype vector.}

\item{y}{The individual-level binary phenotypes with the extremely unbalanced
case-control ratios with dimension n individuals by K phenotypes. It could be
only one phenotype with a n by 1 phenotype vector.}

\item{cov}{The covariate matrix with dimesion n individuals by C covariates.
default: there is no covariates.}
}
\value{
M by K matrix of adjusted test statistics
}
\description{
Calculate the adjusted test statistics for the binary phenotypes
with the extremely unbalanced case-control ratios.
}
\examples{
N <- 10000
M <- 10
y <- replicate(2, sample(c(0,1), N, replace = TRUE, prob = c(0.999, 0.001)))
x <- replicate(M, rbinom(N,2,0.3))
Adjusted_Tstat(x,y)

}
\keyword{Test}
