% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betweenness.R
\name{betweenness}
\alias{betweenness}
\title{The Approximate Betweenness Centrality of Bipartite Network}
\usage{
betweenness(Net)
}
\arguments{
\item{Net}{A adjacency matrix with genotype or genes as rows,
phenotype as columns and the association as entries.
The adjacency matrix of the bipartite network which
could be a simple matrix or data frame.}
}
\value{
A list of object type containing: the first vector is
 is the Approximate Betweenness Centrality of genetic variants,
 the second vector is standardized Approximate Betweenness
 Centralityof genetic variants.
}
\description{
In the bipartite genetic variant and phenotype network (GPN),
the Approximate Betweenness Centrality is used to evaluate
the importance of genetic variant. For instance, the genetic variant
have a high Approximate Betweenness Centrality which means
an important connector between phenotypes.
}
\examples{
Net <- matrix(c(1,1,0,
              1,1,1,
              1,1,1),nrow = 3, ncol = 3, byrow = TRUE)
out <- betweenness(Net)
}
\references{
Latapy, M., Magnien, C., & Del Vecchio, N. (2008).
 \emph{Basic notions for the analysis of large two-mode networks.}
 Social networks, 30(1), 31-48.
}
\keyword{GPN}
