% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_sparsity}
\alias{check_sparsity}
\title{check the adjacency matrix since some nodes have no coonection}
\usage{
check_sparsity(Net)
}
\arguments{
\item{Net}{A adjacency matrix with genotype or genes as rows,
phenotype as columns and the association as entries.
The adjacency matrix of the bipartite network which
could be a simple matrix or data frame.}
}
\description{
check the adjacency matrix since some nodes have no coonection
}
\examples{
Net <- matrix(rnorm(2000)^2, nrow = 500, ncol = 4)
out <- check_sparsity(Net)

}
