% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLC_ave.R
\name{CLC_ave}
\alias{CLC_ave}
\title{Modified Clustering linear combination (CLC) test}
\usage{
CLC_ave(x, y, cov = NULL, cluster.method = "average")
}
\arguments{
\item{x}{The genotyes for only one SNP with dimension n by 1.}

\item{y}{The phentypes with dimension n individuals by K phenotypes.}

\item{cov}{The covariate matrix with dimension n individuals by C covariates. defalut: there is no covariates.}

\item{cluster.method}{The agglomeration method to be used in the hierarchical clustering method. This should be one of "ward.D","ward.D2","single","complete","average","mcquitty","median","centroid". default: "average"}
}
\value{
p-value of the CLC test statistic
}
\description{
CLC test for testing the association between K phenotypes with a SNP. The phenotypes can be either qualitative or binary, espectially the binary phenotypes with the extremely case-control ratio (the test statistics has been adjusted by the saddlepoint approximation)
}
\details{
Reference:
Sha, Q., Wang, Z., Zhang, X., & Zhang, S. (2019). A clustering linear combination approach to jointly analyze multiple phenotypes for GWAS. Bioinformatics, 35(8), 1373-1379.
}
\examples{
N <- 10000
y1 <- replicate(2, sample(c(0,1), N, replace = TRUE, prob = c(0.999, 0.001)))
y2 <- replicate(2, sample(c(0,1), N, replace = TRUE, prob = c(0.5, 0.5)))
y3 <- replicate(2, rnorm(N))
y <- cbind(y1, y2, y3)
x <- rbinom(N,2,0.3)
CLC_ave(x, y)

}
