% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{The cluster coefficient for bipartite network}
\usage{
cluster(Net)
}
\arguments{
\item{Net}{A adjacency matrix with genotype or genes as rows,
phenotype as columns and the association as entries.
The adjacency matrix of the bipartite network which
could be a simple matrix or data frame.}
}
\value{
A list of object type containing: first value
 which is the cluster coefficient of all SNPs/genes,
 the second value is the cluster coefficient of all phenotypes,
 the third value is the cluster coefficient of whole network,
 the last value is the average cluster coefficient of SNPs and
 phenotypes.
}
\description{
The cluster coefficient measures the extent to which
two SNPs share the same set of associated phenotypes.
A high cluster coefficient indicates that SNPs tend
to be associated with similar phenotypes, while
a low cluster coefficient indicates that SNPs
tend to be associated with distinct sets of phenotypes.
The cluster coefficient can be used to identify SNP
sub-networks that are highly interconnected and
potentially functionally related. (It will be time consuming
for high dimensional matrix)
}
\examples{
Net <- matrix(c(1,0,0,1,
              1,1,0,0,
              0,1,1,0),nrow = 4, ncol = 3)
out <- cluster(Net)
}
\keyword{GPN}
