% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommunityDetection2.R
\name{CommunityDetection2}
\alias{CommunityDetection2}
\title{Improved Community Detection method}
\usage{
CommunityDetection2(Net, between_cluster = 0.8)
}
\arguments{
\item{Net}{M by K matrix of the signed bipartite genotype and
phenotype network, row denotes the M SNPs and column denotes K phenotypes.}

\item{between_cluster}{Threshold for hierachical clusters. Default value is 0.8.}
}
\value{
A list of ClusterNumber: the optimal number of clusters (L);
 Cluster: K by L matrix of the optimal cluster results;
 Q_modularity: modularity information.
}
\description{
Improved Community Detection method
}
\examples{
M <- 100
K <- 10
Sigma <- diag(1, nrow = K)
Net <- MASS::mvrnorm(M, rep(0,K), Sigma)
res <- CommunityDetection2(Net, between_cluster = 0.8)
}
