% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectance.R
\name{connectance}
\alias{connectance}
\title{Connectance of bipartie network}
\usage{
connectance(Net)
}
\arguments{
\item{Net}{A adjacency matrix with genotype or genes as rows,
phenotype as columns, and the association as entries.
The adjacency matrix of the bipartite network which
could be a simple matrix or data frame.}
}
\value{
A list of object type containing: a value for
 bipartite network connectance.
}
\description{
Connectance of bipartite network represent the proportion
of realized interactions from the pool of all possible interactions
between two kinds of node in a network.
}
\examples{
\dontrun{
set.seed(123)
Net <- matrix(rnorm(2000)^2, nrow = 500, ncol = 4)
out <- connectance(Net)
}
}
\keyword{GPN}
