% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{degree}
\alias{degree}
\title{Degree of the bipartite network}
\usage{
degree(Net, weight = TRUE)
}
\arguments{
\item{Net}{A adjacency matrix with genotype or genes as rows,
phenotype as columns and the association as entries.
The adjacency matrix of the bipartite network which
could be a simple matrix or data frame.}

\item{weight}{Logical; if TRUE, bipartite projection
will include edge weights. Defaults to TRUE.}
}
\value{
A list of object type containing: node degree
 and standardized node degree
}
\description{
Degree of the bipartite network
}
\examples{
set.seed(123)
Net <- matrix(rnorm(2000)^2, nrow = 500, ncol = 4)
out <- degree(Net)

}
\keyword{GPN}
