% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{degree_entropy}
\alias{degree_entropy}
\alias{cross_entropy}
\alias{kl_divergence}
\title{Bipartite Network Entropy}
\usage{
degree_entropy(Net, weight = TRUE, ...)
}
\arguments{
\item{Net}{A adjacency matrix with genotype or genes as rows,
phenotype as columns and the association as entries.
The adjacency matrix of the bipartite network which
could be a simple matrix or data frame.}

\item{weight}{Logical; if TRUE, bipartite projection
will include edge weights. Defaults to TRUE.}

\item{\dots}{Arguments passed to the internal function
\code{degree_shrink} to avoid no definition of entropy.
The default is \code{ratio = 1e-10}.}

\item{Net.r}{It is a adjacency matrix of the random GPN in which
the degrees of all vertices are the same as the original
bipartite GPN. The adjacency matrix could be
a simple matrix or data frame.}
}
\value{
A list of object type containing: first value
 represents the degree entropy of genetic variants/genes;
 second value represents the degree entropy of phenotype;
 the third value represents the degree entropy of whole
 bipartite network.
}
\description{
Various functions for calculating the Bipartite network entropy
}
\details{
These functions can be used to calculate the entropy of the GPN.

\code{degree_entropy} is one of the Shannon entropy. The Degree entropy
of GPN can be used to measure the diversity of associations
between genetic variants and phenotypes.

\code{cross_entropy} It is used to determine the diversity
between a bipartite GPN and a random network.

\code{kl_divergence} It is used to determine the diversity
between a bipartite GPN and a random network.
}
\examples{
set.seed(123)
Net <- matrix(rnorm(2000)^2, nrow = 500, ncol = 4)
Net.r <- GPN_random(Net)$Net.R
out <- degree_entropy(Net, weight = TRUE)
out2 <- cross_entropy(Net, Net.r)
out3 <- kl_divergence(Net, Net.r)
}
\references{
Kullback, S., & Leibler, R. A. (1951).
 \emph{On information and sufficiency}. The annals of mathematical
 statistics, 22(1), 79-86.

Murphy, K. P. (2012).\emph{ Machine learning:
 a probabilistic perspective}. MIT press.
}
\keyword{GPN}
