% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginv_ratio.R
\name{ginv_ratio}
\alias{ginv_ratio}
\title{Calculate R inverse}
\usage{
ginv_ratio(R, ratio = 1)
}
\arguments{
\item{R}{The squared matrix}

\item{ratio}{The threshold that the smaller eigenvalues to be removed in the calculation of inverse matrix (0 < ratio <=1). default: ratio = 1, that is, the exact inverse matrix; otherwise, ratio < 1 is the adjusted inverse matrix.}
}
\value{
The inverse of matrix R
}
\description{
Calculate the inverse of the squared matrix by either ginv function in MASS package or the proposed method, (Yan et al. 2022, Cao et al. 2022).
}
\examples{

y <- replicate(5, rnorm(100))
R <- cor(y)
ginv_ratio(R)

}
