% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPN.R
\name{GPN}
\alias{GPN}
\title{Construct a signed bipartite genotype and phenotype network}
\usage{
GPN(x, y, cov = NULL, output.T = FALSE)
}
\arguments{
\item{x}{The individual-level genotyes with dimension n individuals by M SNPs.}

\item{y}{The individual-level phentypes with dimesion n individuals by K phenotypes.}

\item{cov}{The covariate matrix with dimesion n individuals by C covariates. defalut: there is no covariates.}

\item{output.T}{Determine if output the test statistics. default: output.T = FALSE}
}
\value{
M by K matrix of the signed bipartite genotype and phenotype network (GPN)
}
\description{
Construct a signed bipartite genotype and phenotype network
}
\examples{
N <- 10000
M <- 10
y1 <- replicate(2, sample(c(0,1), N, replace = TRUE, prob = c(0.999, 0.001))) # two binary phenotypes with the extremely unbalanced case-contorl ratios.
y2 <- replicate(2, sample(c(0,1), N, replace = TRUE, prob = c(0.5, 0.5))) # two binary phenotypes with the balanced case-contorl ratios.
y3 <- replicate(2, rnorm(N)) # two quantitative (continuous) phenotypes
y <- cbind(y1, y2, y3)
x <- replicate(M, rbinom(N,2,0.3))
res <- GPN(x,y)


}
