% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPN_random.R
\name{GPN_random}
\alias{GPN_random}
\title{Generate a random bipartite GPN}
\usage{
GPN_random(Net)
}
\arguments{
\item{Net}{A adjacency matrix with genotype or genes as rows,
phenotype as columns and the association as entries.
The adjacency matrix of the bipartite network which
could be a simple matrix or data frame.}
}
\value{
A list of object type containing: the generated
 random bipartite GPN.
}
\description{
Generates a random bipartite GPN, in which
the degrees of all vertices are the same as the original
bipartite GPN.
}
\examples{
\dontrun{
set.seed(123)
Net <- matrix(rnorm(2000)^2, nrow = 500, ncol = 4)
out <- GPN_random(Net)
}

}
\keyword{GPN}
