% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_data_all.R
\docType{data}
\name{gwas_data_all}
\alias{gwas_data_all}
\title{Example Data: gwas_data_all}
\format{
## `gwas_data_all`
A list with 5 GWAS summary statistics, each have 200 SNPs:
\describe{
  \item{SNP}{SNP name}
  \item{CHR}{Chromsome}
  \item{Beta}{Effect size}
  \item{SE}{Standard error of the effect}
  \item{P}{P-value of the association statistic}
}
}
\usage{
gwas_data_all
}
\description{
This dataset contains an list object from different GWAS summary statistics .
Each list represents a GWAS summary statistics including SNP, CHR, Beta, SE, and P.
}
\keyword{datasets}
