% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_z.R
\docType{data}
\name{gwas_z}
\alias{gwas_z}
\title{Example Data: gwas_z}
\format{
## `gwas_z`
A data frame with 1000 rows and 588 columns:
\describe{
  \item{phenotype}{phenotype index}
  ...
}
}
\usage{
gwas_z
}
\description{
This dataset contains an example data frame. The row
represents the SNPs, and colume is the phenotypes.
}
\keyword{datasets}
