% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCLC.R
\name{HCLC}
\alias{HCLC}
\title{Modified clustering linear combination method based on hierarchical clustering (HCLC)}
\usage{
HCLC(x, y, cov = NULL, cluster.method = "average")
}
\arguments{
\item{x}{The genotyes for only one SNP with dimension n by 1.}

\item{y}{The phentypes with dimension n individuals by K phenotypes.}

\item{cov}{The covariate matrix with dimension n individuals by C covariates. defalut: there is no covariates.}

\item{cluster.method}{The agglomeration method to be used in the hierarchical clustering method. This should be one of "ward.D","ward.D2","single","complete","average","mcquitty","median","centroid". default: "average"}
}
\value{
p-value of the HCLC test statistic
}
\description{
HCLC test for testing the association between K phenotypes with a SNP. The phenotypes can be either qualitative or binary, espectially the binary phenotypes with the extremely case-control ratio (the test statistics has been adjusted by the saddlepoint approximation)
}
\details{
Reference:
Li, X., Zhang, S., & Sha, Q. (2020). Joint analysis of multiple phenotypes using a clustering linear combination method based on hierarchical clustering. Genetic epidemiology, 44(1), 67-78.
Liang, X., Cao, X., Sha, Q., & Zhang, S. (2022). HCLC-FC: a novel statistical method for phenome-wide association studies. bioRxiv.
}
\examples{
N <- 10000
y1 <- replicate(2, sample(c(0,1), N, replace = TRUE, prob = c(0.999, 0.001)))
y2 <- replicate(2, sample(c(0,1), N, replace = TRUE, prob = c(0.5, 0.5)))
y3 <- replicate(2, rnorm(N))
y <- cbind(y1, y2, y3)
x <- rbinom(N,2,0.3)
HCLC(x, y)

}
