% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPhen.R
\name{MultiPhen}
\alias{MultiPhen}
\title{MultiPhen test}
\usage{
MultiPhen(x, y, cov = NULL)
}
\arguments{
\item{x}{The genotyes for only one SNP with dimension n by 1.}

\item{y}{The phentypes with dimension n individuals by K phenotypes.}

\item{cov}{The covariate matrix with dimension n individuals by C covariates. defalut: there is no covariates.}
}
\value{
p-value of the MultiPhen test statistic
}
\description{
MultiPhen test for testing the association between K phenotypes with a SNP.
}
\details{
Reference:
O’Reilly, P. F., Hoggart, C. J., Pomyen, Y., Calboli, F. C., Elliott, P., Jarvelin, M. R., & Coin, L. J. (2012). MultiPhen: joint model of multiple phenotypes can increase discovery in GWAS. PloS one, 7(5), e34861.
}
\examples{
N <- 10000
y1 <- replicate(2, sample(c(0,1), N, replace = TRUE, prob = c(0.999, 0.001)))
y2 <- replicate(2, sample(c(0,1), N, replace = TRUE, prob = c(0.5, 0.5)))
y3 <- replicate(2, rnorm(N))
y <- cbind(y1, y2, y3)
x <- rbinom(N,2,0.3)
MultiPhen(x, y)


}
