% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnibus.R
\name{omnibus}
\alias{omnibus}
\title{Modified omnibus (Chi-square) test}
\usage{
omnibus(x, y, cov = NULL)
}
\arguments{
\item{x}{The genotyes for only one SNP with dimension n by 1.}

\item{y}{The phentypes with dimension n individuals by K phenotypes.}

\item{cov}{The covariate matrix with dimension n individuals by C covariates. defalut: there is no covariates.}
}
\value{
p-value of the omnibus test statistic
}
\description{
Omnibus test for testing the association between K phenotypes with a SNP. The phenotypes can be either qualitative or binary, espectially the binary phenotypes with the extremely case-control ratio (the test statistics has been adjusted by the saddlepoint approximation)
}
\examples{
N <- 10000
y1 <- replicate(2, sample(c(0,1), N, replace = TRUE, prob = c(0.999, 0.001)))
y2 <- replicate(2, sample(c(0,1), N, replace = TRUE, prob = c(0.5, 0.5)))
y3 <- replicate(2, rnorm(N))
y <- cbind(y1, y2, y3)
x <- rbinom(N,2,0.3)
omnibus(x, y)


}
