% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPA_SCORE.R
\name{SPA_SCORE}
\alias{SPA_SCORE}
\title{SPA_SCORE function}
\usage{
SPA_SCORE(x, y, cov = NULL, output.T = FALSE)
}
\arguments{
\item{x}{The individual-level genotyes with dimension n individuals by M SNPs. It could be only one SNPs with a n by 1 genotype vector.}

\item{y}{The individual-level binary phentypes with dimesion n individuals by K phenotypes. It could be only one phenotype with a n by 1 phenotype vector.}

\item{cov}{The covariate matrix with dimesion n individuals by C covariates. defalut: there is no covariates.}

\item{output.T}{Determine if output the test statistics. default: output.T = FALSE}
}
\value{
A list of pvalue: K by M matrix of p-values. If output.T = TRUE, Tstat: K by M matrix of the test statistics.
}
\description{
Score test by saddlepoint approximation for testing the association between only binary phenotypes and SNPs. Note that SPA_SCORE function can use to binary phenotypes with the unbalanced or extremely unbalanced cast-control ratios.
}
\examples{
N <- 10000
M <- 5
K <- 10
x <- replicate(M, rbinom(N,2,0.3))
y <- replicate(K, sample(c(0,1), N, replace = TRUE, prob = c(0.999, 0.001)))
res <- SPA_SCORE(x,y)

}
