% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TATES.R
\name{TATES}
\alias{TATES}
\title{Modified TATES test}
\usage{
TATES(x, y, cov = NULL)
}
\arguments{
\item{x}{The genotyes for only one SNP with dimension n by 1.}

\item{y}{The phentypes with dimension n individuals by K phenotypes.}

\item{cov}{The covariate matrix with dimension n individuals by C covariates. defalut: there is no covariates.}
}
\value{
p-value of the TATES test statistic
}
\description{
TATES test for testing the association between K phenotypes with a SNP. The phenotypes can be either qualitative or binary, espectially the binary phenotypes with the extremely case-control ratio (the test statistics has been adjusted by the saddlepoint approximation)
}
\details{
Reference:
Van der Sluis, S., Posthuma, D., & Dolan, C. V. (2013). TATES: efficient multivariate genotype-phenotype analysis for genome-wide association studies. PLoS genetics, 9(1), e1003235.
}
\examples{
N <- 10000
y1 <- replicate(2, sample(c(0,1), N, replace = TRUE, prob = c(0.999, 0.001)))
y2 <- replicate(2, sample(c(0,1), N, replace = TRUE, prob = c(0.5, 0.5)))
y3 <- replicate(2, rnorm(N))
y <- cbind(y1, y2, y3)
x <- rbinom(N,2,0.3)
TATES(x, y)


}
