% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCORE.R
\name{SCORE}
\alias{SCORE}
\title{SCORE function}
\usage{
SCORE(x, y, cov = NULL)
}
\arguments{
\item{x}{The individual-level genotyes with dimension n individuals by M SNPs. It could be only one SNPs with a n by 1 genotype vector.}

\item{y}{The individual-level phentypes with dimesion n individuals by K phenotypes. It could be only one phenotype with a n by 1 phenotype vector.}

\item{cov}{The covariate matrix with dimesion n individuals by C covariates. defalut: there is no covariates.}
}
\value{
A list of pvalue: K by M matrix of p-values; Tstat: K by M matrix of the test statistics.
}
\description{
Score test for testing the association between phenotypes and SNPs. Note that the phenotypes could be the binary phenotypes with the balanced case-control ratios or quantitative phenotypes. If you consider the phenotypes with the unbalanced or extremely unbalanced cast-control ratios, you need to use SPA_SCORE function.
}
\examples{
N <- 100
M <- 5
K <- 10
x <- replicate(M, rbinom(N,2,0.3))
y <- replicate(K, rnorm(N))
res <- SCORE(x,y)

}
